set openexr location (#14)
PURPOSE: build_devkit.bat script is not working with the latest develop because of changes made in RprUsd repo. This PR fixes this issue.

TECHNICAL DETAILS: In USD repo for ubuntu cmake was forced not to use default search path thus we need to set necessary paths manually. However setting OPENEXR_LOCATION instead of CMAKE_PREFIX_PATH fixed the issue